Fix race condition in integration test (#4994)
* Update command_line_args_test.go

Update command_line_args_test.go

Remove command_line_args_test.go

* Update parameters_test.go

* Update parameters_test.go

* Update parameters_test.go
